requirements.txt
Django~=4.2.11
python-dotenv~=1.0.1
djangorestframework~=3.14.0
mysqlclient~=2.2.4
djangorestframework-simplejwt~=5.3.1
drf-yasg~=1.21.7